+++ src/main/java/kitchenpos/order/handler/OrderMapper.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.util;
+package kitchenpos.order.handler;
 
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
