+++ src/test/java/kitchenpos/manu/application/MenuServiceTest.java
@@ -7,20 +7,17 @@ import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.menu.exception.NotFoundMenuGroupException;
 import kitchenpos.menu.exception.NotFoundProductException;
-import kitchenpos.menu.util.MenuMapper;
-import kitchenpos.menu.util.MenuValidator;
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.menu.handler.MenuMapper;
+import kitchenpos.menu.handler.MenuValidator;
 import kitchenpos.wrap.Price;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.math.BigDecimal;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
