+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -15,8 +15,8 @@ import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.order.exception.EmptyOrderLineItemsException;
 import kitchenpos.order.exception.NotEqualsOrderCountAndMenuCount;
 import kitchenpos.order.exception.NotFoundOrderException;
-import kitchenpos.order.util.OrderMapper;
-import kitchenpos.order.util.OrderValidator;
+import kitchenpos.order.handler.OrderMapper;
+import kitchenpos.order.handler.OrderValidator;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
