+++ src/test/java/kitchenpos/test/application/OrderTableServiceTest.java
@@ -15,8 +15,8 @@ import kitchenpos.table.dto.OrderTableResponse;
 import kitchenpos.table.exception.NotChangeNumberOfGuestThatEmptyTable;
 import kitchenpos.table.exception.NotChangeToEmptyThatCookingOrMealTable;
 import kitchenpos.table.exception.NotFoundOrderTableException;
-import kitchenpos.table.util.OrderTableMapper;
-import kitchenpos.table.util.OrderTableValidator;
+import kitchenpos.table.handler.OrderTableMapper;
+import kitchenpos.table.handler.OrderTableValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
