+++ src/test/java/kitchenpos/test/application/TableGroupServiceTest.java
@@ -13,8 +13,8 @@ import kitchenpos.table.dto.TableGroupResponse;
 import kitchenpos.table.exception.NotChangeToEmptyThatCookingOrMealTable;
 import kitchenpos.table.exception.NotEmptyOrExistTableGroupException;
 import kitchenpos.table.exception.NotFoundOrderTableException;
-import kitchenpos.table.util.TableGroupMapper;
-import kitchenpos.table.util.TableGroupValidator;
+import kitchenpos.table.handler.TableGroupMapper;
+import kitchenpos.table.handler.TableGroupValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
